pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Chawku <chawku.work@gmail.com>"]
readme = "README.md"
packages = [
    {include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^7.1.1"
pytest = "^8.3.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
